move auth out of src/Application.php
